requirements.txt
addict==2.1.1
astroid==1.5.3
autopep8==1.3.2
beautifulsoup4==4.6.0
certifi==2017.7.27.1
chardet==3.0.4
coreapi==2.3.1
coreschema==0.0.4
dj-database-url==0.4.2
Django==1.11.4
django-filter==1.0.4
django-oauth-toolkit==1.0.0
django-rest-swagger==2.1.2
djangorestframework==3.6.4
djangorestframework-jwt==1.11.0
drf-autodocs==0.4.3
drfdocs==0.0.11
gunicorn==19.7.1
idna==2.6
isort==4.2.15
itypes==1.1.0
Jinja2==2.9.6
lazy-object-proxy==1.3.1
Markdown==2.6.9
MarkupSafe==1.0
mccabe==0.6.1
oauthlib==2.0.4
openapi-codec==1.3.2
pep8==1.7.0
psycopg2==2.7.3.1
pycodestyle==2.3.1
PyJWT==1.5.3
pylint==1.7.2
pytz==2017.2
requests==2.18.4
simplejson==3.11.1
six==1.10.0
uritemplate==3.0.0
urllib3==1.22
whitenoise==3.3.1
wrapt==1.10.11
